requirements.txt
addict==2.4.0
ansi2html==1.8.0
appnope==0.1.3
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
Brotli==1.0.9
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.0.4
click==8.1.7
comm==0.1.4
ConfigArgParse==1.7
contourpy==1.1.1
cryptography==41.0.3
cycler==0.12.1
dash==2.14.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.1
filelock==3.13.1
Flask==2.2.5
fonttools==4.43.1
fsspec==2023.10.0
gcloud==0.18.3
gmpy2==2.1.2
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
MarkupSafe==2.1.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
nbformat==5.5.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.26.1
oauth2client==4.1.3
open3d==0.16.1
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.3.1
platformdirs==3.11.0
plotly==5.17.0
prompt-toolkit==3.0.39
protobuf==4.25.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
retrying==1.3.4
rpds-py==0.10.6
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
setuptools==68.2.2
six==1.16.0
stack-data==0.6.2
sympy==1.12
tenacity==8.2.3
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.221
urllib3==1.26.18
wcwidth==0.2.8
Werkzeug==2.2.3
wheel==0.41.2
zipp==3.17.0
